A data set on civilian casualties in civil war; illustrates count regressions

Hultman, Kathman, and Shannon (2013) data set used in McCaskey and Rainey (2015) to illustrate count regression models, such as Poisson, negative binomial, and their zero-inflated and hurdle variants. These are the data to reproduce the negative binomial regression coefficients for Model 1 in Table 1 on p. 884 of Hultman, Kathman, and Shannon (2013).

Usage

hks2013

Format

A country-month data frame.

location

the country

year

the year

mon

the month

osvAll

the number of civilians killed in a conflict month by any combatant faction

troopLag

the number of UN military troops committed to a country in conflict during a given month; lagged one month

policeLag

the number of UN police committed to a country in conflict during a given month; lagged one month

militaryobserversLag

the number of UN observers committed to a country in conflict during a given month; lagged one month

brv_AllLag

the number of all battle deaths; lagged one month

osvAllLagDum

whether any civilians were killed by any combatant faction; lagged one month

incomp

a dichotomous variable that uses the UCDP/PRIO delineation of civil wars fought over territorial (0) or government (1) control.

epduration

the number of months since the beginning of a conflict episode

lntpop

the country's logged population size

For further details, see Hultman, Kathman, and Shannon (2013) pp. 883-884 and Kathman (2013).

References

Hultman, Lisa, Jacob Kathman, and Megan Shannon. 2013. "United Nations Peacekeeping and Civilian Protection in Civil War." American Journal of Political Science 57(4): 875–91. doi:10.1111/ajps.12036 .

Hultman, Lisa, Jacob Kathman, and Megan Shannon. 2013. "Replication data for: United Nations Peacekeeping and Civilian Protection in Civil War." Harvard Dataverse, V3. doi:10.7910/DVN/6EBCGA .

Kathman, Jacob D. 2013. "United Nations Peacekeeping Personnel Commitments, 1990–2011." Conflict Management and Peace Science 30(5): 532–49. doi:10.1177/0738894213491180 .

McCaskey, Kelly, and Carlisle Rainey. 2015. "Substantive Importance and the Veil of Statistical Significance." Statistics, Politics and Policy 6(1–2). doi:10.1515/spp-2015-0001 .

Examples

# a simple example

# load data in a way that mirrors read.csv(), etc.
hks <- crdata::hks2013

# estimate models
f <- osvAll ~ troopLag + policeLag + militaryobserversLag +
  brv_AllLag + osvAllLagDum + incomp + epduration +
  lntpop

# replicates model 1 in table 1 on p. 884 of HKS
# m2 <- MASS::glm.nb(f, data = hks2013,
#                    init.theta = 5,
#                    control = glm.control(epsilon = 1e-12,
#                                          maxit = 2500,
#                                          trace = FALSE))
# summary(m2) # compare to p. 884

# a poisson alternative to the above
m1 <- glm(f, data = hks2013, family = "poisson")
summary(m1)
#> 
#> Call:
#> glm(formula = f, family = "poisson", data = hks2013)
#> 
#> Coefficients:
#>                        Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)          -3.579e+00  4.259e-02  -84.04   <2e-16 ***
#> troopLag             -1.697e-01  1.758e-03  -96.52   <2e-16 ***
#> policeLag            -3.272e+00  2.448e-02 -133.70   <2e-16 ***
#> militaryobserversLag  8.100e+00  1.206e-02  671.54   <2e-16 ***
#> brv_AllLag            5.606e-04  8.476e-06   66.14   <2e-16 ***
#> osvAllLagDum          2.911e-01  4.740e-03   61.42   <2e-16 ***
#> incomp                3.486e+00  1.800e-02  193.69   <2e-16 ***
#> epduration           -2.223e-02  7.940e-05 -279.98   <2e-16 ***
#> lntpop                1.894e-01  2.076e-03   91.24   <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for poisson family taken to be 1)
#> 
#>     Null deviance: 2840157  on 3745  degrees of freedom
#> Residual deviance: 2134801  on 3737  degrees of freedom
#> AIC: 2139137
#> 
#> Number of Fisher Scoring iterations: 11
#>